Step over nested elements for now
For now, keep things simple & don't evaluate nested matches.

If we want to add support for nesting later, then we'd probably want some kind
of depth-first evaluation, with the expansion of an inner element passed to
outer elements. This is quite complex, so wait until we actually have a need
for this.