Drop support for node < 4
We use Object.assign, which is part of ES6. The stable node releases are all
on 4 now, and larger projects with a need for 0.10 compatibility generally
pull in shims like `core-js`.

On the client, the main use of `elematch` would be in a ServiceWorker, which
in turn means that `Object.assign` is natively supported.