Add streaming support
In order to support streaming content transformations, we need to be able to
separate fully processed content from an incomplete remainder section. Our
focus is on matching elements, so we need to make sure that we split the input
between fully parsed elements & a partial-tag remainder.

For now, the need to parse complete tags is tricky to combine with support for
the (minor) isXML optimization. The win from that optimization were fairly
minor in any case, so this patch just drops support for this optimization
internally.

In order to support returning matches and a remainder, the return value is
changed to an object:

```javascript
{
  matches: ['<div>first</div>', '<div>second</div>'],
  remainder: '<div'
}
```

Performance seems to be fairly unchanged overall.